Fix for not wrapping different wrappee.
In cases, like in OpenCL, the main library (OpenCL) acts as a wrapper for other implementations (libamdocl64.so). In this case both have same function names implemented. If we wrap OpenCL functions with gotcha which dynamically loads libamdocl64.so with same function name, we should not return a wrapper meant to intercept OpenCL.
So we should check if the wrapper given to the caller of dlsym is for the same wrappee as requested by the caller.